lib/agentlib: move TA user management to agent library
