+++ lib/agentlib/meson.build
@@ -64,6 +64,7 @@ sources += files(
     'process.c',
     'symbols.c',
     'unistd.c',
+    'users.c',
 )
 te_libs += [
     'tools',
